Made it Work
Comics can now be loaded to the page from the box (must have http:// in
front). This will not work with all comics. Further extractor training
will have to be done to correct this. The urls are stored in local
storage, so they will reload onto the page even if the tab is closed.